+++ test/rhsmlib_test/facts/test_cloud_facts.py
