+++ test/rhsmlib_test/facts/test_collector.py
