+++ test/rhsmlib_test/facts/test_host_collector.py
