+++ test/rhsmlib_test/facts/test_kpatch.py
