+++ test/rhsmlib_test/facts/test_pkg_arches.py
