+++ test/rhsmlib_test/services/test_config.py
@@ -13,15 +13,9 @@
 #
 import unittest
 
-import dbus
-
 from rhsm.config import RhsmConfigParser, NoOptionError
-from rhsmlib.dbus import constants
-from rhsmlib.dbus.objects.config import ConfigDBusObject
 from rhsmlib.services.config import Config, ConfigSection
-from test.rhsmlib_test.base import DBusObjectTest, TestUtilsMixin
-
-from test import subman_marker_dbus
+from test.rhsmlib_test.base import TestUtilsMixin
 
 TEST_CONFIG = """
 [foo]
@@ -177,54 +171,3 @@ class TestConfigSection(BaseConfigTest):
     def test_in(self):
         self.assertIn("quux", self.config["foo"])
         self.assertNotIn("missing", self.config["foo"])
-
-
-@subman_marker_dbus
-class TestConfigDBusObject(DBusObjectTest, TestUtilsMixin):
-    def setUp(self):
-        super(TestConfigDBusObject, self).setUp()
-        self.proxy = self.proxy_for(ConfigDBusObject.default_dbus_path)
-        self.interface = dbus.Interface(self.proxy, constants.CONFIG_INTERFACE)
-
-    def dbus_objects(self):
-        self.fid = self.write_temp_file(TEST_CONFIG)
-        self.addCleanup(self.fid.close)
-        self.parser = RhsmConfigParser(self.fid.name)
-        return [(ConfigDBusObject, {"parser": self.parser})]
-
-    def test_get_all(self):
-        def assertions(*args):
-            result = args[0]
-            self.assertIn("server", result)
-
-        dbus_method_args = [""]
-        self.dbus_request(assertions, self.interface.GetAll, dbus_method_args)
-
-    def test_get_property(self):
-        def assertions(*args):
-            result = args[0]
-            self.assertIn("server.example.com", result)
-
-        dbus_method_args = ["server.hostname", ""]
-        self.dbus_request(assertions, self.interface.Get, dbus_method_args)
-
-    def test_get_section(self):
-        def assertions(*args):
-            result = args[0]
-            self.assertIn("hostname", result)
-
-        dbus_method_args = ["server", ""]
-        self.dbus_request(assertions, self.interface.Get, dbus_method_args)
-
-    def test_set(self):
-        def assertions(*args):
-            self.assertEqual("new", self.parser.get("server", "hostname"))
-
-        dbus_method_args = ["server.hostname", "new", ""]
-        self.dbus_request(assertions, self.interface.Set, dbus_method_args)
-
-    def test_set_section_fails(self):
-        dbus_method_args = ["server", "new", ""]
-
-        with self.assertRaisesRegex(dbus.DBusException, r"Setting an entire section is not.*"):
-            self.dbus_request(None, self.interface.Set, dbus_method_args)
